refactor out StripeConnectWebView
